Get accurate token counts via non-stream API call
- After streaming response completes, make non-stream API call to get
  real token usage from VLLM instead of using estimated (len/4)
- Fix percentage display to use float for proper decimal display (0.8% vs 0%)
- Now showing accurate input+output tokens like: 1,491 (0.8% / 196,608)

+++ tui/main.go
@@ -2176,8 +2176,9 @@ func (m model) View() string {
 	if m.totalContextTokens > 0 {
 		var ctxStr string
 		if m.contextWindow > 0 {
-			pct := int((m.totalContextTokens * 100) / m.contextWindow)
-			ctxStr = fmt.Sprintf("\033[90m%s (%d%% / %s)\033[0m", formatTokenCount(m.totalContextTokens), pct, formatTokenCount(m.contextWindow))
+			// Use float for proper rounding
+			pct := float64(m.totalContextTokens) * 100 / float64(m.contextWindow)
+			ctxStr = fmt.Sprintf("\033[90m%s (%.1f%% / %s)\033[0m", formatTokenCount(m.totalContextTokens), pct, formatTokenCount(m.contextWindow))
 		} else {
 			ctxStr = fmt.Sprintf("\033[90m%s\033[0m", formatTokenCount(m.totalContextTokens))
 		}
